CapsuleDiceMovement
Made the capsule be able to move to dice I'm pretty sure. Just need to make it move in smaller increments and rotate 90 degrees every 10 steps so it stays on the board. On Monday I'm gonna start making tiles so i know the distance that the capsule has to start moving around the board.